Split out loginUserWithSecurityKey
